Adds the Teleport command, written by Alvin Alexander.
Teleport is an improved version of the traditional Unix/Linux/GNU cd
command. It wraps the cd command in order to maintain a history of
visited directories which can be displayed in order to move, teleport,
between them using either their history entry number or with partial
directory names. For basic use, tp is just like cd.