Hook direnv in bash
direnv will load/unload environment variables based on $PWD

Before each prompt, direnv checks for a `.envrc` file in the current
and parent directories. If the file exists (and is authorized),
it is loaded into a bash sub-shell and all exported variables are then
captured by direnv and then made available to the current shell.